fix logic for --openssl-config in NODE_OPTIONS; rejigger hapi compat checking and add to the set-framework-*.js test files
