+++ test/instrumentation/modules/hapi/set-framework-hapi.js
@@ -2,26 +2,27 @@
 
 const agent = require('../../../..').start({
   captureExceptions: true,
-  metricsInterval: 0
+  metricsInterval: 0,
+  centralConfig: false
 })
 
-const pkg = require('@hapi/hapi/package')
-const semver = require('semver')
-
-// hapi 17+ requires Node.js 8.9.0 or higher
-if (semver.lt(process.version, '8.9.0') && semver.gte(pkg.version, '17.0.0')) process.exit()
+var isHapiIncompat = require('../../../_is_hapi_incompat')
+if (isHapiIncompat('hapi')) {
+  // Skip out of this test.
+  process.exit()
+}
 
 let asserts = 0
 
 agent.setFramework = function ({ name, version, overwrite }) {
   asserts++
   assert.strictEqual(name, 'hapi')
-  assert.strictEqual(version, require('@hapi/hapi/package').version)
+  assert.strictEqual(version, require('hapi/package').version)
   assert.strictEqual(overwrite, false)
 }
 
 const assert = require('assert')
 
-require('@hapi/hapi')
+require('hapi')
 
 assert.strictEqual(asserts, 1)
